+++ eugene_shurpach/diploma/conftest.py
