+++ eugene_shurpach/diploma/tools/base_page.py
