+++ eugene_shurpach/diploma/tools/constants.py
