+++ eugene_shurpach/diploma/tools/duck_page.py
@@ -1,6 +1,6 @@
 from selenium.webdriver import Keys
-from eugene_shurpach.Final_task.tools.base_page import BasePage
-from eugene_shurpach.Final_task.tools.duck_page_locators import DuckPageLocators
+from eugene_shurpach.diploma.tools.base_page import BasePage
+from eugene_shurpach.diploma.tools.duck_page_locators import DuckPageLocators
 
 
 class DuckPage(BasePage):
