+++ eugene_shurpach/diploma/tools/duck_page_locators.py
