+++ eugene_shurpach/diploma/tools/ducks_page_locators.py
