+++ eugene_shurpach/diploma/tools/main_page.py
@@ -1,8 +1,8 @@
-from eugene_shurpach.Final_task.tools.base_page import BasePage
-from eugene_shurpach.Final_task.tools.basket_page import BasketPage
-from eugene_shurpach.Final_task.tools.constants import USER_EMAIL, USER_PASS
-from eugene_shurpach.Final_task.tools.ducks_page import DucksPage
-from eugene_shurpach.Final_task.tools.main_page_locators import MainPageLocators
+from eugene_shurpach.diploma.tools.base_page import BasePage
+from eugene_shurpach.diploma.tools.basket_page import BasketPage
+from eugene_shurpach.diploma.tools.constants import USER_EMAIL, USER_PASS
+from eugene_shurpach.diploma.tools.ducks_page import DucksPage
+from eugene_shurpach.diploma.tools.main_page_locators import MainPageLocators
 
 
 class MainPage(BasePage):
