+++ eugene_shurpach/diploma/tools/main_page_locators.py
