+++ eugene_shurpach/diploma/tools/success_page_locators.py
